[BOJ]1로 만들기(1463)/실버3
